LayoutTests/imported/w3c/web-platform-tests/navigation-timing/test_readwrite-expected.txt

Description

This test validates that the window.performance object is read/write.


PASS window.performance is defined
PASS window.performance is read/write
PASS var performance is read/write
